tests/integration/requirements.txt
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
pytest-mock>=3.7.0
confluent-kafka>=1.8.2
websockets>=10.1
psycopg2-binary>=2.9.3
PyJWT>=2.3.0
redis>=4.0.0
aioredis>=2.0.0
docker>=5.0.0
requests>=2.25.0
sqlalchemy>=1.4.0
passlib>=1.7.4
argon2-cffi>=21.3.0
qrcode>=7.3.0
pillow>=9.0.0
python-dotenv>=0.19.0
flask-sqlalchemy>=3.0.0